--- src/api/user/user-name/index.ts
@@ -1,6 +1,0 @@
-import { api } from '@/api/api';
-
-export const getUserName = async () => {
-  const response = await api.get('/user/profile');
-  return response.data;
-};
