+++ src/pages/onboarding/completion/index.tsx
@@ -2,18 +2,12 @@ import { textBox } from './index.styles';
 import { Complete } from '@/assets/icons';
 import { useNavigate } from 'react-router-dom';
 import { Text, Button } from '@/components';
-import { useState, useEffect } from 'react';
-import { useGetUserNameQuery } from '@/queries/user/user-name';
+import { useGetUserNameQuery } from '@/queries/user';
 import theme from '@/styles/theme';
 
 export default function Completion() {
   const navigate = useNavigate();
   const { data } = useGetUserNameQuery();
-  const [nickName, setNickName] = useState('');
-
-  useEffect(() => {
-    setNickName(data.nickname);
-  }, []);
   const handleCompleteButtonClick = () => {
     navigate('/', { replace: true });
   };
@@ -22,7 +16,7 @@ export default function Completion() {
     <div css={textBox}>
       <Complete width="70px" />
       <Text typo="title1" color={theme.colors.green200}>
-        {nickName} 님<br />
+        {data.nickname} 님<br />
         <Text typo="title1" color={theme.colors.black}>
           회원가입이 완료됐어요!
         </Text>
