+++ src/queries/products/index.ts
@@ -1,7 +1,7 @@
 import { getProducts } from '@/api/products';
 import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEYS } from '../query-keys';
-import { GetProductsParamType } from '@/types';
+import { GetProductsParamType } from '@/types/product';
 
 export const useGetProductsQuery = (params: GetProductsParamType) => {
   return useQuery({
