+++ src/types/product/index.ts
@@ -14,18 +14,3 @@ export interface ProductType {
   sellerId: string;
   status: string;
 }
-
-export interface PatchUserOnboardingBody {
-  email: string;
-  nickname: string;
-  phone: string;
-}
-
-export interface PostCheckNicknameBody {
-  nickname: string;
-}
-
-export interface PostUserPetBody {
-  name: string;
-  petType: string;
-}
