+++ src/types/user/index.ts
@@ -1,0 +1,14 @@
+export interface PatchUserOnboardingBody {
+  email: string;
+  nickname: string;
+  phone: string;
+}
+
+export interface PostCheckNicknameBody {
+  nickname: string;
+}
+
+export interface PostUserPetBody {
+  name: string;
+  petType: string;
+}
